test d'envois depuis VS

+++ TestOpenStack/Nova.cs
@@ -9,7 +9,7 @@ namespace TestOpenStack
 {
     public partial class Nova : Form
     {
-        private ComputeService compute = new ComputeService(connect.GetConnection(),"regionOne",true);
+        private ComputeService compute = new ComputeService(connect.GetConnection(),"regionOne");
         private NetworkingService networking = new NetworkingService(connect.GetConnection(), "regionOne");
 
         public Nova()
@@ -20,33 +20,36 @@ namespace TestOpenStack
 
         public async void listVM()
         {
-            DGV_listVM.Rows.Clear();
+            //DGV_listVM.Rows.Clear();
             IPage<Server> listServer = await compute.ListServersAsync();
-            foreach(Server server in listServer)
+            DataGridView newDGV = DGV_listVM;
+            newDGV.Rows.Clear();
+            foreach (Server server in listServer)
             {
-                DGV_listVM.Rows.Add(server.Id, server.Name);
+                newDGV.Rows.Add(server.Id, server.Name);
+                //DGV_listVM.Rows.Add(server.Id, server.Name);
             }
+            DGV_listVM = newDGV;
         }
 
         private async void B_suppr_VM_Click(object sender, EventArgs e)
         {
             /* message de confirmation */
+            DialogResult result = MessageBox.Show("Attention, vous etes sur le point de supprimer une instance d'un serveur." + '\n' + "Voulez vous vraiment la supprimer ?", "Demande de suppression d'instance virtuelle", MessageBoxButtons.YesNo);
 
-            /* fin message de confirmation */
+            if (result == DialogResult.Yes)
+            {
+                Server server = await compute.GetServerAsync((Identifier)DGV_listVM.SelectedRows[0].Cells[0].Value);
+                await server.DeleteAsync();
+            }
 
-            Server server = await compute.GetServerAsync((Identifier)DGV_listVM.SelectedRows[0].Cells[0].Value);
-            await server.DeleteAsync();
+            TBar_deverouille.Value = 0;
         }
 
         private void B_ajout_VM_Click(object sender, EventArgs e)
         {
             preCreationInstance PCI = new preCreationInstance();
             PCI.Visible = true;
-        }
-
-        private void B_modif_VM_Click(object sender, EventArgs e)
-        {
-
         }
 
         private void active_suppression(object sender, EventArgs e)
@@ -59,11 +62,6 @@ namespace TestOpenStack
             {
                 B_suppr_VM.Enabled = false;
             }
-        }
-
-        private void Nova_FormClosed(object sender, FormClosedEventArgs e)
-        {
-
         }
 
         private async void B_VM_Start_Click(object sender, EventArgs e)
@@ -110,48 +108,121 @@ namespace TestOpenStack
             await server.RebootAsync();
         }
 
-        private void button1_Click(object sender, EventArgs e)
-        {
-
-        }
-
         private async void selectVM(object sender, DataGridViewCellEventArgs e)
         {
             if (DGV_listVM.SelectedRows != null)
             {
                 //on demande au serveur de nous renvoyer la liste des serveur suivant les options recherche spécifié
                 Server server = await compute.GetServerAsync((Identifier)DGV_listVM.SelectedRows[0].Cells[0].Value);
-
                 //MessageBox.Show("nom: " + server.Name + " // etat server: " + server.Status);
-
-                IList<ServerVolumeReference> listVolumeReference = server.AttachedVolumes;
-                foreach (ServerVolumeReference volumeReference in listVolumeReference)
+                try
                 {
-                    Volume V = await compute.GetVolumeAsync(volumeReference.Id);
-                    LB_volume.Items.Add(V.Name);
+                    IList<ServerVolumeReference> listVolumeReference = server.AttachedVolumes;
+                    foreach (ServerVolumeReference volumeReference in listVolumeReference)
+                    {
+                        LB_volume.Items.Clear();
+                        Volume V = await compute.GetVolumeAsync(volumeReference.Id);
+                        LB_volume.Items.Add(V.Name);
+                    }
+                }
+                catch (Exception ex){}
+
+                try{
+                    L_created.Text = server.Created.Value.ToString();
+                }catch (Exception ex)
+                {
+                    L_created.Text = "Non renseigné";
                 }
 
-                L_created.Text = server.Created.Value.ToString();
-                L_configDisk.Text = server.DiskConfig.DisplayName;
+                try{
+                    L_configDisk.Text = server.DiskConfig.DisplayName;
+                }catch (Exception ex)
+                {
+                    L_configDisk.Text = "Non renseigné";
+                }
 
                 //récupérer le flavor
-                Flavor F = await compute.GetFlavorAsync(server.Flavor.Id);
-                L_flavor.Text = F.Name;
-                L_id.Text = server.Id.ToString();
+                try
+                {
+                    Flavor F = await compute.GetFlavorAsync(server.Flavor.Id);
+                    L_flavor.Text = F.Name;
+                }catch(Exception ex)
+                {
+                    L_flavor.Text = "Non rensigné";
+                }
+
+                try
+                {
+                    L_id.Text = server.Id.ToString();
+                }catch(Exception ex)
+                {
+                    L_id.Text = "Non rensigné";
+                }
 
                 //récupérer l'image
-                Image I = await compute.GetImageAsync(server.Image.Id);
-                L_image.Text = I.Name;
-                L_IPV4.Text = server.IPv4Address;
-                L_IPV6.Text = server.IPv6Address;
-                L_nom.Text = server.Name;
-                L_progress.Text = server.Progress.ToString();
-                if (!(server.Progress > 0))
+                try
+                {
+                    Image I = await compute.GetImageAsync(server.Image.Id);
+                    L_image.Text = I.Name;
+                }catch(Exception ex)
+                {
+                    L_image.Text = "Non renseigné";
+                }
+
+                try
+                {
+                    L_IPV4.Text = server.IPv4Address;
+                }catch(Exception ex)
+                {
+                    L_IPV4.Text = "Non renseigné";
+                }
+
+                try
+                {
+                    L_IPV6.Text = server.IPv6Address;
+                }catch(Exception ex)
+                {
+                    L_IPV6.Text = "Non rensigné";
+                }
+
+                try
+                {
+                    L_nom.Text = server.Name;
+                }catch(Exception ex)
+                {
+                    L_nom.Text = "Non renseigné";
+                }
+
+                try
+                {
+                    L_progress.Text = server.Progress.ToString();
+                }catch(Exception ex)
+                {
+                    L_progress.Text = "Non renseigné";
+                }
+
+                try
                 {
                     L_connected.Text = server.Status.DisplayName;
-                    //L_lastExec.Text = server.Launched.Value.ToString();
+                }catch(Exception ex)
+                {
+                    L_connected.Text = "Non renseigné";
                 }
-                consoleInterface(server);
+
+                try
+                {
+                    L_lastExec.Text = server.Launched.Value.ToString();
+                }catch(Exception ex)
+                {
+                    L_lastExec.Text = "Non rensigné";
+                }
+
+                try
+                {
+                    consoleInterface(server);
+                }catch(Exception ex){}
+
+                TBar_deverouille.Value = 0;
             }
         }
 
@@ -164,8 +235,11 @@ namespace TestOpenStack
         {
             if (DGV_listVM.Rows.Count > 0)
             {
-                Server server = await compute.GetServerAsync((Identifier)DGV_listVM.SelectedRows[0].Cells[0].Value);
-                consoleInterface(server);
+                try
+                {
+                    Server server = await compute.GetServerAsync((Identifier)DGV_listVM.SelectedRows[0].Cells[0].Value);
+                    consoleInterface(server);
+                }catch(Exception ex){}
             }
         }
     }
